fix: @web3-storage/access/agent no longer exports authorizeWithPollClaim (#656)
Motivation:
* https://github.com/web3-storage/w3protocol/issues/633

Note:
* this was an unintentional addition to our package exports and not an
intentional addition to our public API, so this PR conventional commit
message is not marked as a breaking api change
* I notified the only person I expect to have relied on it
https://github.com/web3-storage/w3protocol/pull/602#issuecomment-1482000952
